Support overloaded types in C++ templates
On 32bit systems both int and long are typically both 32bit, while
long long is 64bit, but both are distinct types. On 64bit systems the
same is true for long and long long (both 64bit). But as they are
distinct types, only one of them is aliased to the corresponding sized
type (int32_t/int64_t). This means that for example on 64bit Linux
systems, where long is typically aliased to int64_t, it is not possible
to create a libdivide::divider<long long>. Conversely, on Apple macOS
int64_t is typically long long, so it's not possible to create a
libdivide::divider<long> there - even though the types are compatible
with each other.

This commit adds a bit of template magic that tries to detect the
proper sized type for any given type libdivide is used with. This
allows libdivide to support divider<int>, divider<long> and
divider<long long> on all architectures where these map to either 32bit
or 64bit types.

The code should be compatible with C++98.