Reimplement vectorized high-multiplies
The previous implementations were taken from StackOverflow, but the
licensing there is ambiguous. Revert back to the previous implementation
which was bespoke and so covered by the libdivide license, and use it in
all vector regimes.

One new idea relative to the old implementation is to use a shuffle
instead of a shift when moving the high 32 bits of packed 64 bit values
into lower lanes. The advantage of a shuffle is that it preserves the
source register which saves us a mov.